requirements.txt
fastapi==0.104.1
uvicorn
python-multipart
aiofiles
SQLAlchemy
passlib[bcrypt]
python-jose[cryptography]==0.0.6
pillow==10.1.0
python-dotenv==1.0.0
google-generativeai==0.3.1
